Tests passing.
Had to move a few tests     to the "TestsToBeFixed" as they appear to be using some sort of tablequalification
that either doesn't work at all, or bigquery doesn't accept it anymore.

Requires setting up src/test/resources with the following properties files:
+ serviceaccount.properties
+ installedaccount.properties
+ installedaccount1.properties

Added a sample properties file.

Along with any p12 files required for these credentials.

Also requires uploading the efashion dataset somewhere. I found this in some SAP HANA tutorials
I will think about the best ways to add this data so that one can add it to their project.

Removed a significant amount of System.err logging.

There are also tests that seemto be passing even though they should probably be failing. These tests point
to the original starschema__clouddb project/dataset which I don't have access to.

Added License files to resources so that they will be added to the JAR as well as to the pom file.

Finally, Added an ApplicationName (Starschema BigQuery JDBC Driver) to the Driver to stop a warning.