Fixup the way the sync query code handles cancelling a job by reference (#54)
Cancel code relied on existence of a job on BQStatement, but that was only set (only could be set) by the async code. We need to wait for the sync code to time out, then cancel the query. Reduces the timeout on the sync API to only five seconds to enable faster cancellation.