Remove support for Catalog -> ProjectId name mangling
A prior version of this driver supported translating between SQL-92 and
BigQuery Legacy dialects.  As part of this support, it would mangle
BigQuery Project Ids like `looker.com:sample-project` to
`looker_com__sample-project`.  We found that there were some bugs in
this translation leading to the mangled name being sent to BigQuery when
the real ProjectId should have been sent.

We recently removed support for this translation between dialects.
Since we no longer require that, we can now remove most of the code
which did the name mangling.

In this PR, we fix on using the Project Id internally as much as
possible.  Only in public methods of the JDBC API do we accept a Catalog
name.  All internal methods translate this to a Project Id before doing
any further processing of them.

We do retain one vestige of the original name mangling.  We continue to
support the mangled names in the JDBC URI.  This is to maintain
compatibility with an existing configurations which used the mangled
name.  Should we decide to drop this support in the future, it should be
trivial now that all of the mangling is contained in the single
`CatalogName.toProjectId` method.