Disable pretty printing by default
According to feedback from the BigQuery team there is significant
overhead in the whitespace processing for BigQuery pretty printed JSON.

They recommend disabling it.

Unfortunately, doesn't seem to be a clean way to set it by default, so
I've had to search every execute() call and add it beforehand.

They call out bigquery.jobs.getQueryResults and bigquery.tabledata.list
as the big winners.