Return Unauthorized when using invalid OAuth token
Updating the google auth library changed the behavior of handling an
invalid oauth token.  Previously, it would return a failure response as
Unauthorized.  With the change, an interaction in the google libraries
would cause it to throw an IllegalStateException.

This change smooths over that interaction so that we can return an
Unauthorized response as we did before.