Prevent NPE if connection is closed before result set is initialized  (#125)
Edge case was causing a NPE to be thrown when initializing BQForwardOnlyResultSet. The underlying issue seems to be a race condition where a connection is closed after the job has been completed but before the driver has an opportunity to get the query results. Specifically, calling close() on a BQConnection sets its bigquery attribute to null so efforts to getQueryResultsDivided would fail.

This just adds an additional check to see if bigquery is null and throws a more informative exception.