feat: set quota project when using impersonation (#131)
When using SA impersonation we need to explicitly set a quota project that will be used for billing services. Here we add the projectId specified in the JDBC URL as the quota project when constructing the ImpersonatedCredentials to use. This effectively makes the projectId declared in the URL the billing project that will be used for all BQ jobs.