Prepare for v3.2.0 release (#202)
This CL bumps the minor version of the BQJDBC driver. In a recently
merged PR[0], we added the ability to configure a default dataset
project id, which is used in conjunction with the dataset id to locate
unqualified table references. We added this new capability in a
backwards compatible manner.

[0] https://github.com/looker-open-source/bqjdbc/pull/201